requirements.txt
openai~=1.7.2
fastapi==0.100.1
pydantic-settings==2.0.3
requests==2.31.0
ujson==5.8.0
uvicorn==0.22.0
python-dotenv~=1.0.0
pydantic~=2.0.3